tests/wpt/mozilla/tests/mozilla/document-characterSet-invalid.html

<!DOCTYPE html>
<meta charset="unknown-charset">
<html>
    <head>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script>
            test(function() {
                assert_equals(document.characterSet, "UTF-8", "test1-0, unknown charset");
            });
        </script>
    </head>
</html>
